+++ nts.uk/web/nts.uk.pr.web/src/main/webapp/view/base/simplehistory/updatehistory/viewmodel.ts
@@ -112,11 +112,14 @@ module nts.uk.pr.view.base.simplehistory.updatehistory {
                     startYearMonth: self.startYearMonth()
                 };
                 if (self.actionType() == ScreenModel.ACTION_TYPE_DELETE) {
-                      self.dialogOptions.onDeleteCallBack(callBackData);
+                    nts.uk.ui.dialog.confirm("データを削除します。\r\n よろしいですか？").ifYes(function() {
+                        self.dialogOptions.onDeleteCallBack(callBackData);
+                        nts.uk.ui.windows.close();
+                    });
                 } else {
                     self.dialogOptions.onUpdateCallBack(callBackData);
+                    nts.uk.ui.windows.close();
                 }
-                nts.uk.ui.windows.close();
             }
 
             /**
